Fix inverted SVG arc sweep (#294)
Fixes #276

Replaces the SVG paths in the `svg` example with some off-the-shelf
icons that users might expect to render correctly.

With this change to `svg.rs`, they now render correctly.

This is a breaking change, so anyone relying on the old behavior will
need to invert the arc sweeps in their SVGs.